+++ source/.packages.yaml
@@ -1,7 +1,7 @@
 '@handbook/code-block':
-  version: 1.0.0-alpha.4
+  version: 1.0.0-alpha.5
   tag: latest
   module: esm
 '@handbook/*':
-  version: 1.0.0-alpha.4
+  version: 1.0.0-alpha.5
   tag: latest
