+++ source/package.json
@@ -34,6 +34,7 @@
     "@types/node-fetch": "^2.5.7",
     "@types/prettier": "^2.1.0",
     "@types/react": "^16.9.48",
+    "@types/yargs": "^15.0.5",
     "fs-extra": "^9.0.1",
     "jest": "^26.3.0",
     "lint-staged": "^10.2.13",
@@ -54,6 +55,7 @@
     "simple-git": "^2.20.1",
     "typescript": "^4.0.2",
     "unified": "^9.2.0",
+    "yargs": "^15.4.1",
     "yargs-parser": "^19.0.4"
   }
 }
