+++ source/src/@handbook/remark-magic-comments/index.ts
@@ -1,8 +1,7 @@
+import { isRootNode, Node } from '@handbook/remark-node-types';
 import { Plugin } from 'unified';
-import { Node } from 'unist';
 import { processCommand } from './processCommand';
 import { transformCommand } from './transformCommand';
-import { isRootNode } from './types';
 
 const plugin: Plugin<[]> = () => (node, { dirname = process.cwd() }): Promise<Node> | void => {
   if (isRootNode(node)) {
