+++ source/src/@handbook/remark-split-html-lines/index.ts
@@ -1,7 +1,6 @@
+import { isRootNode, Node } from '@handbook/remark-node-types';
 import { Plugin } from 'unified';
-import { Node } from 'unist';
 import { transform } from './transform';
-import { isRootNode } from './types';
 
 const plugin: Plugin<[]> = () => (node): Promise<Node> | void => {
   if (isRootNode(node)) {
